Add webpack and lint configuration
