+++ penny-guess-client/.eslintignore
@@ -1,0 +1,3 @@
+target
+node
+node_nodules
