+++ penny-guess-client/.eslintrc.json
@@ -1,0 +1,6 @@
+{
+  "extends": "airbnb",
+  "globals": {
+    "fetch": false
+  }
+}
