+++ penny-guess-client/package.json
@@ -8,14 +8,13 @@
   "author": "Andrew Barnes",
   "license": "ISC",
   "scripts": {
-    "test": "mocha --compilers js:babel-core/register --require ./src/test/test_helper.js 'src/test/**/*.@(js)' --recursive",
-    "test:watch": "mocha --compilers js:babel-core/register --require ./src/test/test_helper.js 'src/test/**/*.@(js)' --recursive --watch",
+    "lint": "eslint --ext .js,.jsx .",
+    "lint:fix": "eslint --ext .js,.jsx src --fix",
+    "test": "mocha --compilers js:babel-./register --require ./src/test/test_helper.js 'src/test/**/*.@(js)' --recursive",
     "mock": "cd src/mock && node server.js",
-    "start": "webpack-dev-server --inline --hot",
-    "prebuild": "rm -rf ./target/* &&  npm install --no-optional",
-    "build": "webpack --mode development",
-    "prebuildprod": "npm run prebuild",
-    "buildprod": "webpack --mode production"
+    "start": "webpack-dev-server --inline --hot --config webpack.dev.js",
+    "prebuild": "rimraf ./target/classes/META-INF/resources/*",
+    "build": "webpack --config webpack.prod.js"
   },
   "dependencies": {
     "@babel/runtime": "^7.4.3",
@@ -33,12 +32,22 @@
     "babel-loader": "^8.0.5",
     "chalk": "^2.4.2",
     "css-loader": "^2.1.1",
+    "eslint": "^5.3.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-react": "^7.12.4",
+    "exlint": "^0.1.6",
     "fs": "0.0.1-security",
     "html-webpack-plugin": "^3.2.0",
+    "jshint": "^2.10.2",
     "json-server": "^0.14.2",
+    "rimraf": "^2.6.3",
     "style-loader": "^0.23.1",
+    "uglifyjs-webpack-plugin": "^2.1.2",
     "webpack": "^4.29.6",
     "webpack-cli": "^3.3.0",
-    "webpack-dev-server": "^3.3.1"
+    "webpack-dev-server": "^3.3.1",
+    "webpack-merge": "^4.2.1"
   }
 }
