+++ penny-guess-client/webpack.prod.js
@@ -1,0 +1,16 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common');
+const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
+
+module.exports = merge(common, {
+  mode: 'production',
+  devtool: 'source-map',
+  optimization: {
+    minimizer: [
+      new UglifyJsPlugin({
+        sourceMap: true,
+        parallel: true,
+      })
+    ]
+  }
+});
