+++ penny-guess-server/src/main/resources/public/index.html
@@ -16,6 +16,6 @@
     Please enable javascript to view this website
   </div>
 </noscript>
-<script type="text/javascript" src="/webjars/penny-guess-client/1.0.0/main.js"></script>
+<script type="text/javascript" src="/penny-guess-client-1.0.0.js"></script>
 </body>
 </html>
